Components
Seperated the logic for the settings modal into seperate components for easier maintanence and, testing and reusability.
